frontend/requirements.txt
dash
dash-bootstrap-components
dash-auth-external
requests
python-dotenv
pandas
plotly
gunicorn
